chore(query-orchestrator): Extract Pre-aggregation-related classes to separate files (#9274)
* some typos and refactoring

* rafactor CacheKey type def

* typo

* refactor imports in time.ts

* just typo

* remove unused: redisPrefix from PreAggregationLoadCache class

* remove unused: redisPrefix from PreAggregationLoader & PreAggregationPartitionRangeLoader

* refactor PreAggregations: move every class to a separate file

* refactor PreAggregations: delete moved

* code reformat

* better naming in utcToLocalTimeZone()

* code polishment in PreAggregationLoadCache class

* a bit of code polishment in PreAggregationLoader class

* attempt to fix CacheKey type
