+++ packages/cubejs-query-orchestrator/src/orchestrator/QueryCache.ts
@@ -96,16 +96,16 @@ export type TempTable = {
  * definition (stored in the second element) link.
  */
 export type PreAggTableToTempTable = [
-  string, // common table name (without sufix)
+  string, // common table name (without suffix)
   TempTable,
 ];
 
+export type CacheKeyItem = string | string[] | QueryTuple | QueryTuple[] | undefined;
+
 export type CacheKey =
-  | string
-  | [
-      query: string | QueryTuple,
-      options?: string[]
-    ];
+  [CacheKeyItem, CacheKeyItem] |
+  [CacheKeyItem, CacheKeyItem, CacheKeyItem] |
+  [CacheKeyItem, CacheKeyItem, CacheKeyItem, CacheKeyItem];
 
 type CacheEntry = {
   time: number;
@@ -187,7 +187,7 @@ export class QueryCache {
   /**
    * Generates from the `queryBody` the final `sql` query and push it to
    * the queue. Returns promise which will be resolved by the different
-   * objects, depend from the original `queryBody` object. For the
+   * objects, depend on the original `queryBody` object. For the
    * persistent queries returns the `stream.Writable` instance.
    *
    * @throw Error
@@ -374,7 +374,7 @@ export class QueryCache {
   }
 
   public static queryCacheKey(queryBody: QueryBody): CacheKey {
-    const key = [
+    const key: CacheKey = [
       queryBody.query,
       queryBody.values,
       (queryBody.preAggregations || []).map(p => p.loadSql)
@@ -384,7 +384,7 @@ export class QueryCache {
     }
     // @ts-ignore
     key.persistent = queryBody.persistent;
-    return <CacheKey>key;
+    return key;
   }
 
   protected static replaceAll(replaceThis, withThis, inThis) {
