+++ packages/cubejs-query-orchestrator/src/orchestrator/index.ts
@@ -3,6 +3,9 @@ export * from './ContinueWaitError';
 export * from './LocalCacheDriver';
 export * from './LocalQueueDriver';
 export * from './PreAggregations';
+export * from './PreAggregationPartitionRangeLoader';
+export * from './PreAggregationLoader';
+export * from './PreAggregationLoadCache';
 export * from './QueryCache';
 export * from './QueryOrchestrator';
 export * from './QueryQueue';
